+++ src/_playground/documentation/Separator/Separator.js
@@ -1,12 +1,7 @@
 import React from 'react';
 
 export const Separator = () => {
-    const separatorStyle = {
-        display: 'block',
-        margin: '2rem 0',
-        borderBottom: '1px solid #E3E3E3'
-    };
     return (
-        <div style={separatorStyle} />
+        <div className='frDocs-Content__separator' />
     );
 };